+++ app/internal/service/service.go
@@ -6,8 +6,8 @@ import (
 )
 
 type UserService interface {
-	Get(ctx context.Context, id int) (model.User, error)
-	Create(ctx context.Context, user model.UserCreate) (int, error)
+	Get(ctx context.Context, id uint64) (model.User, error)
+	Create(ctx context.Context, user model.UserCreate) (uint64, error)
 	Update(ctx context.Context, user model.UserUpdate) error
-	Delete(ctx context.Context, id int) error
+	Delete(ctx context.Context, id uint64) error
 }
